feat: 添加消息表情反应后端功能 + 前端修复/补全
后端:
- 新增 DB 迁移 014_message_reactions (message_reactions 表)
- 新增 MessageReaction Model / ReactionService / Handler
- 注册路由 /api/v1/messages/:id/reactions (GET/POST/DELETE)
- WebSocket 协议增加 AddReaction/RemoveReaction/ReactionAdded/ReactionRemoved
- WebSocket handler 实现反应添加/移除 + 房间广播
- MessageResponse 增加 reactions 字段 (批量加载, N+1 优化)
- 更新 AppState / mod.rs / routes 注册

前端 (user_v2):
- 修复 useMessageActions roomId 过期导致发送失败
- 修复 selectRoom/selectDirectRoom 缺少 JoinRoom WS 消息
- 修复 http.ts 拦截器 /auth/login 401 被吞
- 修复 friend API 路径缺少 /users 前缀
- 修复 getMyRooms 返回类型 / settings API 路径
- 修复 storeToRefs 渲染崩溃改为 computed
- FriendsView / DiscoverView / RoomSettingsModal 完整实现
- SearchMessagesPanel / 浏览器推送通知 / 消息搜索键盘导航
- 内联 SVG 统一替换为 Element Plus 图标
- 修复 typingUsers 模板绑定 .value / store export 遗漏

文档:
- 更新 docs/api/v1/http/messages.md (reaction 端点 + 示例)
- 更新 docs/api/v1/websocket/message.md (WS 反应事件)
- 更新 docs/api/v1/features-summary.md (反应功能特性)
- 更新 docs/api/index.md / websocket/index.md
- 更新 client/user_v2/docs/roadmap.md

+++ src/services/room_service.rs
@@ -1595,7 +1595,7 @@ impl RoomService {
         // 人均房间数
         let avg_rooms_per_user: f64 = sqlx::query_scalar(
             r#"
-            SELECT COALESCE(AVG(room_count), 0.0)
+            SELECT COALESCE(AVG(room_count)::float8, 0.0)
             FROM (
                 SELECT user_id, COUNT(*) as room_count
                 FROM room_members
@@ -1609,7 +1609,7 @@ impl RoomService {
         // 平均成员数
         let avg_members_per_room: f64 = sqlx::query_scalar(
             r#"
-            SELECT COALESCE(AVG(member_count), 0.0)
+            SELECT COALESCE(AVG(member_count)::float8, 0.0)
             FROM (
                 SELECT room_id, COUNT(*) as member_count
                 FROM room_members
